Configure release orchestration, step 1 of n (#288)
This change adds .release/ci.hcl. This file is "read by the
crt-orchestrator app and should include the events required
to trigger the common workflows responsible for processing the
artifacts built in the previous phase."

For "step 1 of n," we simply add the common `prepare` workflow,
which crt-orchestrator will run after a successful `build` workflow
run. The only side effect of `prepare` is upload to HashiCorp's
artifact store. That keeps this change a safe distance away from
existing staging and production releases.